backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
email-validator
jinja2
python-multipart
